backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Queue
redis==5.0.1
rq==1.16.1

# Storage
boto3==1.34.23
minio==7.2.3

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Image Processing
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# AI/LLM
google-generativeai==0.3.2
google-cloud-vision==3.7.0

# Utilities
python-dateutil==2.8.2
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.26.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Monitoring & Logging
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Data validation
pydantic[email]==2.5.3

# File handling
filetype==1.2.0
